requirements.txt
# Core dependencies
python-dotenv==1.1.1
typer[all]==0.16.0
rich==14.0.0

# Notion integration
notion-client==2.4.0

# Data validation
pydantic==2.11.7
pydantic_core==2.33.2

# HTTP requests
requests==2.32.4
beautifulsoup4==4.13.4
soupsieve==2.7

# AI/LLM integration
openai==1.95.1

# Web automation
playwright==1.53.0
pyee==13.0.0
greenlet==3.2.3

# CLI tools
click==8.2.1
shellingham==1.5.4

# Template rendering
jinja2==3.1.6
MarkupSafe==3.0.2

# Resume validator dependencies
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
attrs==25.3.0
referencing==0.36.2
rpds-py==0.26.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0

# Logging
structlog==25.4.0

# Development tools
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0

# Additional dependencies
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.2
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
